+++ src/base/StandardLoggingHandler.h
@@ -32,7 +32,7 @@ public:
     virtual ~StandardLoggingHandler ();
 
     virtual void logMessage(const tm* pTime, unsigned millis, const string& category,
-            long level, const UTF8String& sMsg);
+            unsigned level, const UTF8String& sMsg);
 };
 
 }
